Task 3
I modified the code to generate the test cases for JamesBond. I changed
the runFSM method to type boolean, that checks for if "007" is in the
outputPattern variable. My diagram was constructed with Dia, a .png is
also saved. I did not successfully implement JamesBond using regular
expressions, only contains.